+++ src/toolbar/list-button.tsx
@@ -1,5 +1,6 @@
-import { type PointerEvent, useCallback, useId } from "react";
-import { useFocused, useSlate } from "slate-react";
+import { useCallback, useId } from "react";
+import type { CallbackEvent } from "@shopify/app-bridge-ui-types";
+import { ReactEditor, useFocused, useSlate } from "slate-react";
 import { type ListElement } from "~/types";
 import { isListActive, toggleList } from "~/helper/list";
 
@@ -13,9 +14,10 @@ export function ListButton({ listType }: ListButtonProps) {
   const isActive = isListActive(editor, listType) && focused;
   const tooltipId = useId();
 
-  const handleOnClick = useCallback((event: PointerEvent<HTMLSpanElement>) => {
+  const handleOnClick = useCallback((event: CallbackEvent<'s-button'>) => {
     event.preventDefault();
     toggleList(editor, listType);
+    ReactEditor.focus(editor);
   }, [editor, listType]);
 
   return (