+++ vite.config.ts
@@ -3,6 +3,14 @@ import { defineConfig } from 'vite';
 
 export default defineConfig({
   esbuild: {
-    target: "es2024"
-  }
+    target: "es2024",
+  },
+  optimizeDeps: {
+    exclude: ['@shopify/app-bridge-ui-types'],
+  },
+  build: {
+    rollupOptions: {
+      external: ['@shopify/app-bridge-ui-types'],
+    },
+  },
 });